доп ограничение по юзеру

+++ main/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth.models import User
 from django.contrib.auth.decorators import permission_required
 from . models import Services, Clients, Orders
 
+main_user = "popov"
 
 def index(request):
     all_services = Services.objects.all()
@@ -15,52 +16,61 @@ def index(request):
 
 @permission_required('main.view_services')
 def administration_service(request):
-    all_services = Services.objects.all()
+    if request.user.username == main_user:
+        all_services = Services.objects.all()
 
-    if request.method == "POST":
-        new_name = request.POST.get('service_name')
-        new_description = request.POST.get('service_description')
-        new_price = request.POST.get('service_price')
+        if request.method == "POST":
+            new_name = request.POST.get('service_name')
+            new_description = request.POST.get('service_description')
+            new_price = request.POST.get('service_price')
 
-        new_service = Services(
-            name=new_name,
-            description=new_description,
-            price=new_price
-        )
-        new_service.save()
-        return redirect("administration_service")
+            new_service = Services(
+                name=new_name,
+                description=new_description,
+                price=new_price
+            )
+            new_service.save()
+            return redirect("administration_service")
 
 
-    return render(request, 'main/administration_service.html', {
-        'title': 'Услуги',
-        'all_services': all_services,
-    })
-
+        return render(request, 'main/administration_service.html', {
+            'title': 'Услуги',
+            'all_services': all_services,
+        })
+    else:
+        return redirect('main')
 
 def edit_service(request, id):
-    service = Services.objects.get(id=id)
-    if request.method == 'POST':
-        new_name = request.POST.get('service_name')
-        new_description = request.POST.get('service_description')
-        new_price = request.POST.get('service_price')
-        service.name = new_name
-        service.description = new_description
-        service.price = new_price
-        service.save()
+    if request.user.username == main_user:
+        service = Services.objects.get(id=id)
+        if request.method == 'POST':
+            new_name = request.POST.get('service_name')
+            new_description = request.POST.get('service_description')
+            new_price = request.POST.get('service_price')
+            service.name = new_name
+            service.description = new_description
+            service.price = new_price
+            service.save()
 
-        return redirect('administration_service')
+            return redirect('administration_service')
 
+        return render(request, 'main/administration_service_edit.html', {
+            'title': service.name,
+            'service': service
+        })
+    else:
+        return redirect('main')
 
-    return render(request, 'main/administration_service_edit.html', {
-        'title': service.name,
-        'service': service
-    })
 
 
 def delete_service(request, id):
-    service = Services.objects.get(id=id)
-    service.delete()
-    return redirect('administration_service')
+    if request.user.username == main_user:
+        service = Services.objects.get(id=id)
+        service.delete()
+        return redirect('administration_service')
+    else:
+        return redirect('main')
+
